requirements.txt
bcrypt==4.0.1
bitcoin==1.1.39
blinker==1.6.2
blockcypher==1.0.93
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2
